+++ test/integration/queue/create_message_test.rb
@@ -35,7 +35,12 @@ describe Azure::Storage::Queue::QueueService do
 
     it "creates a message in the specified queue and returns nil on success" do
       result = subject.create_message(queue_name, message_text)
-      result.must_be_nil
+      result.wont_be_nil
+      result.wont_be_empty
+      result.length.must_equal 1
+      result[0].message_text.must_be_nil
+      result[0].pop_receipt.wont_be_nil
+      result[0].id.wont_be_nil
 
       result = subject.peek_messages queue_name
       result.wont_be_nil
@@ -50,7 +55,12 @@ describe Azure::Storage::Queue::QueueService do
 
       it "the :visibility_timeout option causes the message to be invisible for a period of time" do
         result = subject.create_message(queue_name, message_text, visibility_timeout: visibility_timeout)
-        result.must_be_nil
+        result.wont_be_nil
+        result.wont_be_empty
+        result.length.must_equal 1
+        result[0].message_text.must_be_nil
+        result[0].pop_receipt.wont_be_nil
+        result[0].id.wont_be_nil
 
         result = subject.peek_messages queue_name
         result.length.must_equal 0
@@ -64,7 +74,12 @@ describe Azure::Storage::Queue::QueueService do
 
       it "the :message_ttl option modifies the expiration_date of the message" do
         result = subject.create_message(queue_name, message_text, message_ttl: message_ttl)
-        result.must_be_nil
+        result.wont_be_nil
+        result.wont_be_empty
+        result.length.must_equal 1
+        result[0].message_text.must_be_nil
+        result[0].pop_receipt.wont_be_nil
+        result[0].id.wont_be_nil
 
         result = subject.peek_messages queue_name
         result.wont_be_nil